Move modifier into ui package
